refactor: split characterization suite under the 600-line file cap
